@@ -11,16 +11,14 @@
     "license": "ISC",
     "type": "module",
     "dependencies": {
-        "@discordjs/voice": "^0.16.1",
-        "discord.js": "^14.14.1",
-        "dotenv": "^16.3.1",
-        "sodium-native": "^4.0.4",
-        "typescript": "^5.3.3",
-        "usetube": "^2.2.7",
-        "youtube-sr": "^4.3.10",
-        "ytdl-core": "^4.11.5"
+        "@discordjs/voice": "^0.19.0",
+        "@snazzah/davey": "^0.1.9",
+        "discord.js": "^14.25.1",
+        "dotenv": "^16.6.1",
+        "typescript": "^5.9.3",
+        "youtubei.js": "^16.0.1"
     },
     "devDependencies": {
-        "@tsconfig/node20": "^20.1.2"
+        "@tsconfig/node20": "^20.1.8"
     }
 }
